+++ kakao/src/main/kotlin/com/agoda/kakao/actions/BottomNavigationViewActions.kt
@@ -1,0 +1,31 @@
+package com.agoda.kakao.actions
+
+import android.support.design.widget.BottomNavigationView
+import android.support.test.espresso.UiController
+import android.support.test.espresso.ViewAction
+import android.support.test.espresso.matcher.ViewMatchers
+import android.view.View
+
+/**
+ * Provides actions for BottomNavigationView
+ */
+interface BottomNavigationViewActions : BaseActions {
+    /**
+     * Sets the given item id as selected
+     *
+     * @param id menu item id to be set
+     */
+    fun setSelectedItem(id: Int) {
+        view.perform(object : ViewAction {
+            override fun getDescription() = "Sets given item id as selected: $id"
+
+            override fun getConstraints() = ViewMatchers.isAssignableFrom(BottomNavigationView::class.java)
+
+            override fun perform(uiController: UiController, view: View) {
+                if (view is BottomNavigationView) {
+                    view.selectedItemId = id
+                }
+            }
+        })
+    }
+}
