+++ kakao/src/main/kotlin/com/agoda/kakao/actions/EditableActions.kt
@@ -1,0 +1,33 @@
+package com.agoda.kakao.actions
+
+import android.support.test.espresso.action.ViewActions
+
+/**
+ * Provides editable actions for views
+ */
+interface EditableActions : BaseActions {
+    /**
+     * Types the given text into the view
+     *
+     * @param text Text to input
+     */
+    fun typeText(text: String) {
+        view.perform(ViewActions.typeText(text))
+    }
+
+    /**
+     * Replaces the current view text with given
+     *
+     * @param text Text to input instead of current
+     */
+    fun replaceText(text: String) {
+        view.perform(ViewActions.replaceText(text))
+    }
+
+    /**
+     * Clears current text in the view
+     */
+    fun clearText() {
+        replaceText("")
+    }
+}
