+++ kakao/src/main/kotlin/com/agoda/kakao/actions/RatingBarActions.kt
@@ -1,0 +1,31 @@
+package com.agoda.kakao.actions
+
+import android.support.test.espresso.UiController
+import android.support.test.espresso.ViewAction
+import android.support.test.espresso.matcher.ViewMatchers
+import android.view.View
+import android.widget.RatingBar
+
+/**
+ * Provides action for RatingBar
+ */
+interface RatingBarActions : BaseActions {
+    /**
+     * Set rating for RatingBar
+     *
+     * @param number rating to set for the RatingBar
+     */
+    fun setRatingAt(number: Float) {
+        view.perform(object : ViewAction {
+            override fun getDescription() = "set rating value of rating bar as: $number"
+
+            override fun getConstraints() = ViewMatchers.isAssignableFrom(RatingBar::class.java)
+
+            override fun perform(uiController: UiController, view: View) {
+                if (view is RatingBar) {
+                    view.rating = number
+                }
+            }
+        })
+    }
+}
